+++ extensions/vscode/src/extension/VsCodeExtension.ts
@@ -32,7 +32,7 @@ import { Battery } from "../util/battery";
 import { TabAutocompleteModel } from "../util/loadAutocompleteModel";
 import type { VsCodeWebviewProtocol } from "../webviewProtocol";
 import { VsCodeMessenger } from "./VsCodeMessenger";
-import { startAiderProcess } from "../integrations/aider/aider";
+import { startAiderProcess } from "../integrations/aider/aiderUtil";
 
 export class VsCodeExtension {
   // Currently some of these are public so they can be used in testing (test/test-suites)
