+++ extensions/vscode/src/integrations/aider/aiderUtil.ts
@@ -3,7 +3,7 @@ import * as cp from "child_process";
 import { Core } from "core/core";
 import { ContinueGUIWebviewViewProvider } from "../../ContinueGUIWebviewViewProvider";
 import { getIntegrationTab } from "../../util/integrationUtils";
-import Aider from "core/llm/llms/Aider";
+import Aider from "core/llm/llms/AiderLLM";
 import { execSync } from "child_process";
 import { isFirstPearAICreatorLaunch } from "../../copySettings";
 
@@ -212,7 +212,7 @@ async function checkPythonInstallation(): Promise<boolean> {
   return false;
 }
 
-async function checkAiderInstallation(): Promise<boolean> {
+export async function checkAiderInstallation(): Promise<boolean> {
   const commands = [
     "aider --version",
     "python -m aider --version",
